Initial attempt to fix the build. Main issue is the missing 'packaging' section the pom.xml
