+++ src/main/java/org/killbill/billing/plugin/payment/dao/gen/DefaultCatalog.java
@@ -1,7 +1,7 @@
 /*
  * This file is generated by jOOQ.
  */
-package org.killbill.billing.plugin.paymenttest.dao.gen;
+package org.killbill.billing.plugin.payment.dao.gen;
 
 
 import java.util.ArrayList;
@@ -37,7 +37,7 @@ public class DefaultCatalog extends CatalogImpl {
     /**
      * The schema <code>killbill</code>.
      */
-    public final Killbill KILLBILL = org.killbill.billing.plugin.paymenttest.dao.gen.Killbill.KILLBILL;
+    public final Killbill KILLBILL = org.killbill.billing.plugin.payment.dao.gen.Killbill.KILLBILL;
 
     /**
      * No further instances allowed
