+++ src/main/java/org/killbill/billing/plugin/payment/dao/gen/Keys.java
@@ -1,7 +1,7 @@
 /*
  * This file is generated by jOOQ.
  */
-package org.killbill.billing.plugin.paymenttest.dao.gen;
+package org.killbill.billing.plugin.payment.dao.gen;
 
 
 import javax.annotation.Generated;
@@ -10,12 +10,12 @@ import org.jooq.Identity;
 import org.jooq.UniqueKey;
 import org.jooq.impl.AbstractKeys;
 import org.jooq.types.ULong;
-import org.killbill.billing.plugin.paymenttest.dao.gen.tables.TestpaymentHppRequests;
-import org.killbill.billing.plugin.paymenttest.dao.gen.tables.TestpaymentPaymentMethods;
-import org.killbill.billing.plugin.paymenttest.dao.gen.tables.TestpaymentResponses;
-import org.killbill.billing.plugin.paymenttest.dao.gen.tables.records.TestpaymentHppRequestsRecord;
-import org.killbill.billing.plugin.paymenttest.dao.gen.tables.records.TestpaymentPaymentMethodsRecord;
-import org.killbill.billing.plugin.paymenttest.dao.gen.tables.records.TestpaymentResponsesRecord;
+import org.killbill.billing.plugin.payment.dao.gen.tables.TestpaymentHppRequests;
+import org.killbill.billing.plugin.payment.dao.gen.tables.TestpaymentPaymentMethods;
+import org.killbill.billing.plugin.payment.dao.gen.tables.TestpaymentResponses;
+import org.killbill.billing.plugin.payment.dao.gen.tables.records.TestpaymentHppRequestsRecord;
+import org.killbill.billing.plugin.payment.dao.gen.tables.records.TestpaymentPaymentMethodsRecord;
+import org.killbill.billing.plugin.payment.dao.gen.tables.records.TestpaymentResponsesRecord;
 
 
 /**
