+++ src/main/java/org/killbill/billing/plugin/payment/dao/gen/Killbill.java
@@ -1,7 +1,7 @@
 /*
  * This file is generated by jOOQ.
  */
-package org.killbill.billing.plugin.paymenttest.dao.gen;
+package org.killbill.billing.plugin.payment.dao.gen;
 
 
 import java.util.ArrayList;
@@ -13,9 +13,9 @@ import javax.annotation.Generated;
 import org.jooq.Catalog;
 import org.jooq.Table;
 import org.jooq.impl.SchemaImpl;
-import org.killbill.billing.plugin.paymenttest.dao.gen.tables.TestpaymentHppRequests;
-import org.killbill.billing.plugin.paymenttest.dao.gen.tables.TestpaymentPaymentMethods;
-import org.killbill.billing.plugin.paymenttest.dao.gen.tables.TestpaymentResponses;
+import org.killbill.billing.plugin.payment.dao.gen.tables.TestpaymentHppRequests;
+import org.killbill.billing.plugin.payment.dao.gen.tables.TestpaymentPaymentMethods;
+import org.killbill.billing.plugin.payment.dao.gen.tables.TestpaymentResponses;
 
 
 /**
@@ -41,17 +41,17 @@ public class Killbill extends SchemaImpl {
     /**
      * The table <code>killbill.testpayment_hpp_requests</code>.
      */
-    public final TestpaymentHppRequests TESTPAYMENT_HPP_REQUESTS = org.killbill.billing.plugin.paymenttest.dao.gen.tables.TestpaymentHppRequests.TESTPAYMENT_HPP_REQUESTS;
+    public final TestpaymentHppRequests TESTPAYMENT_HPP_REQUESTS = org.killbill.billing.plugin.payment.dao.gen.tables.TestpaymentHppRequests.TESTPAYMENT_HPP_REQUESTS;
 
     /**
      * The table <code>killbill.testpayment_payment_methods</code>.
      */
-    public final TestpaymentPaymentMethods TESTPAYMENT_PAYMENT_METHODS = org.killbill.billing.plugin.paymenttest.dao.gen.tables.TestpaymentPaymentMethods.TESTPAYMENT_PAYMENT_METHODS;
+    public final TestpaymentPaymentMethods TESTPAYMENT_PAYMENT_METHODS = org.killbill.billing.plugin.payment.dao.gen.tables.TestpaymentPaymentMethods.TESTPAYMENT_PAYMENT_METHODS;
 
     /**
      * The table <code>killbill.testpayment_responses</code>.
      */
-    public final TestpaymentResponses TESTPAYMENT_RESPONSES = org.killbill.billing.plugin.paymenttest.dao.gen.tables.TestpaymentResponses.TESTPAYMENT_RESPONSES;
+    public final TestpaymentResponses TESTPAYMENT_RESPONSES = org.killbill.billing.plugin.payment.dao.gen.tables.TestpaymentResponses.TESTPAYMENT_RESPONSES;
 
     /**
      * No further instances allowed
