+++ src/main/java/org/killbill/billing/plugin/payment/dao/gen/Tables.java
@@ -1,0 +1,41 @@
+/*
+ * This file is generated by jOOQ.
+ */
+package org.killbill.billing.plugin.payment.dao.gen;
+
+
+import javax.annotation.Generated;
+
+import org.killbill.billing.plugin.payment.dao.gen.tables.TestpaymentHppRequests;
+import org.killbill.billing.plugin.payment.dao.gen.tables.TestpaymentPaymentMethods;
+import org.killbill.billing.plugin.payment.dao.gen.tables.TestpaymentResponses;
+
+
+/**
+ * Convenience access to all tables in killbill
+ */
+@Generated(
+        value = {
+                "http://www.jooq.org",
+                "jOOQ version:3.9.1"
+        },
+        comments = "This class is generated by jOOQ"
+)
+@SuppressWarnings({"all", "unchecked", "rawtypes"})
+public class Tables {
+
+    /**
+     * The table <code>killbill.testpayment_hpp_requests</code>.
+     */
+    public static final TestpaymentHppRequests TESTPAYMENT_HPP_REQUESTS = org.killbill.billing.plugin.payment.dao.gen.tables.TestpaymentHppRequests.TESTPAYMENT_HPP_REQUESTS;
+
+    /**
+     * The table <code>killbill.testpayment_payment_methods</code>.
+     */
+    public static final TestpaymentPaymentMethods TESTPAYMENT_PAYMENT_METHODS = org.killbill.billing.plugin.payment.dao.gen.tables.TestpaymentPaymentMethods.TESTPAYMENT_PAYMENT_METHODS;
+
+    /**
+     * The table <code>killbill.testpayment_responses</code>.
+     */
+    public static final TestpaymentResponses TESTPAYMENT_RESPONSES = org.killbill.billing.plugin.payment.dao.gen.tables.TestpaymentResponses.TESTPAYMENT_RESPONSES;
+}
