+++ src/main/java/org/killbill/billing/plugin/payment/dao/gen/tables/TestpaymentResponses.java
@@ -1,7 +1,7 @@
 /*
  * This file is generated by jOOQ.
  */
-package org.killbill.billing.plugin.paymenttest.dao.gen.tables;
+package org.killbill.billing.plugin.payment.dao.gen.tables;
 
 
 import java.math.BigDecimal;
@@ -19,9 +19,9 @@ import org.jooq.TableField;
 import org.jooq.UniqueKey;
 import org.jooq.impl.TableImpl;
 import org.jooq.types.ULong;
-import org.killbill.billing.plugin.paymenttest.dao.gen.Keys;
-import org.killbill.billing.plugin.paymenttest.dao.gen.Killbill;
-import org.killbill.billing.plugin.paymenttest.dao.gen.tables.records.TestpaymentResponsesRecord;
+import org.killbill.billing.plugin.payment.dao.gen.Keys;
+import org.killbill.billing.plugin.payment.dao.gen.Killbill;
+import org.killbill.billing.plugin.payment.dao.gen.tables.records.TestpaymentResponsesRecord;
 
 
 /**
