+++ src/main/java/org/killbill/billing/plugin/payment/dao/gen/tables/records/TestpaymentPaymentMethodsRecord.java
@@ -1,7 +1,7 @@
 /*
  * This file is generated by jOOQ.
  */
-package org.killbill.billing.plugin.paymenttest.dao.gen.tables.records;
+package org.killbill.billing.plugin.payment.dao.gen.tables.records;
 
 
 import java.sql.Timestamp;
@@ -14,7 +14,7 @@ import org.jooq.Record8;
 import org.jooq.Row8;
 import org.jooq.impl.UpdatableRecordImpl;
 import org.jooq.types.ULong;
-import org.killbill.billing.plugin.paymenttest.dao.gen.tables.TestpaymentPaymentMethods;
+import org.killbill.billing.plugin.payment.dao.gen.tables.TestpaymentPaymentMethods;
 
 
 /**
