+++ src/main/java/org/killbill/billing/plugin/payment/dao/gen/tables/records/TestpaymentResponsesRecord.java
@@ -1,7 +1,7 @@
 /*
  * This file is generated by jOOQ.
  */
-package org.killbill.billing.plugin.paymenttest.dao.gen.tables.records;
+package org.killbill.billing.plugin.payment.dao.gen.tables.records;
 
 
 import java.math.BigDecimal;
@@ -15,7 +15,7 @@ import org.jooq.Record10;
 import org.jooq.Row10;
 import org.jooq.impl.UpdatableRecordImpl;
 import org.jooq.types.ULong;
-import org.killbill.billing.plugin.paymenttest.dao.gen.tables.TestpaymentResponses;
+import org.killbill.billing.plugin.payment.dao.gen.tables.TestpaymentResponses;
 
 
 /**
