+++ src/test/java/org/killbill/billing/plugin/payment/model/PayloadTest.java
@@ -1,4 +1,4 @@
-package org.killbill.billing.plugin.paymenttest.model;
+package org.killbill.billing.plugin.payment.model;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
